feat(a 60fps): Added a 60fps animation demo with particle effects and mouse interact...
# Project State

## Current Focus
Added a 60fps animation demo with particle effects and mouse interaction in the game loop example

## Context
This implements a visual showcase for the compositor's animation capabilities, demonstrating smooth rendering at 60fps with interactive particle effects triggered by mouse clicks

## Completed
- [x] Implemented particle system with physics (gravity, velocity)
- [x] Added mouse interaction for particle bursts
- [x] Created starfield background with twinkling effects
- [x] Implemented rocket animation with trail particles
- [x] Added turbo mode toggle (space key)
- [x] Included help display toggle (? key)
- [x] Set up proper terminal initialization/cleanup
- [x] Implemented FPS counter and timing control

## In Progress
- [ ] No active work in progress

## Blockers
- None identified

## Next Steps
1. Add sound effects for particle bursts
2. Implement score system based on click count
3. Add level progression or difficulty scaling
---
category: feat
scope: a 60fps
---

+++ src/framework/widgets/kanban.rs
@@ -670,7 +670,7 @@ impl crate::framework::widget::Widget for Kanban {
             }
             crate::input::event::MouseEventKind::ScrollDown => {
                 let visible = (area.width / (self.column_width + 1)) as usize;
-                if self.scroll_offset + visible < self.columns.len() {
+                if (self.scroll_offset as usize) + visible < self.columns.len() {
                     self.scroll_offset += 1;
                     self.dirty = true;
                 }
